Fix error related to throttling `ResendConfirmationEmail`
By duplicating the request as we are accessing request.DATA twice. To get the
request DATA, `djangorestframework` `read` the `POST` request (a stream)
without rewinding the stream.
This was producing an `Assertion` error happenign only in a test where
we were calling the `view` with the same `request` twices returning:
{'detail': 'JSON parse error - No JSON object could be decoded'}`
https://github.com/tomchristie/django-rest-framework/blob/650a91ac24cbd3e5b4ad5d7d7c6706fdf6160a78/rest_framework/parsers.py#L60-L64